+++ op/init.go
@@ -1,0 +1,9 @@
+package op
+
+import (
+	"github.com/viant/afs"
+)
+
+func init() {
+	afs.GetRegistry().Register(Scheme, Provider)
+}